--- src/src/js/localizations/resources/components/loginWithTfa.ts
@@ -1,18 +1,0 @@
-export default {
-    en: {
-        title: "Login With Two Factor Authentication",
-        message: {
-            authenticatorRequired: "The Authenticator code field is required."
-        },
-        field: {
-            twoFactorInfo: "Your login is protected with an authenticator app. Enter your authenticator code below.",
-            twoFactorCode: "Authenticator Code",
-            recoveryInfo: "Don't have access to your authenticator device?",
-            recoveryCode: "You can log in with a recovery code",
-            rememberThisMachine: "Remember this machine"
-        },
-        button: {
-            signIn: "Sign in"
-        }
-    }
-};
