+++ src/src/js/localizations/resources/components/signInWithRecoveryCode.ts
